chore(DTFS2-7431): resolved comments on PR
