+++ e2e-tests/portal/cypress/support/portal/completeAboutBuyerSection.js
@@ -1,0 +1,11 @@
+const { fillBuyerDetails } = require('./createBssEwcsDeal');
+/**
+ * Completes the 'About Buyer' section of a form.
+ *
+ * @description Calls the `fillBuyerDetails` function to populate the section with buyer details.
+ */
+const completeAboutBuyerSection = () => {
+  fillBuyerDetails();
+};
+
+export { completeAboutBuyerSection };
